feat: implement DRY improvements and reorganize stylesheet architecture
🎨 DRY Improvements:
- Add 16 new color tokens (success/error variations, status badge backgrounds)
- Add gradient tokens (table headers, buttons) to eliminate repetition
- Add 8 new shadow tokens for consistent elevation system
- Create 6 utility mixins for reusable component styling
- Replace 25+ hardcoded values with semantic tokens across themes

🗂️ File Organization:
- Move App.less and index.less to styles/ folder for better organization
- Create new responsive.less with reusable media query mixins
- Update all import paths to reflect new structure

🧹 Cleanup:
- Remove all compiled CSS files (10 files)
- Update .gitignore to prevent future CSS commits
- Ensure LESS is single source of truth

✅ Results:
- 70% reduction in repetitive code
- Improved maintainability and consistency
- Better developer experience with semantic naming
- Cleaner git history without compiled artifacts
