+++ apps/ratio-ui/src/styles/App.less
@@ -3,8 +3,9 @@
 // =========================
 
 // Import design tokens and mixins
-@import 'styles/tokens/_index.less';
-@import 'styles/mixins/artist.less';
+@import 'tokens/_index.less';
+@import 'mixins/artist.less';
+@import 'mixins/responsive.less';
 
 // CSS Layers for better cascade management
 @layer reset, tokens, base, components, utilities, themes, animations;
