+++ apps/ratio-ui/src/styles/index.less
@@ -3,7 +3,7 @@
 // =========================
 
 // Import design tokens
-@import 'styles/tokens/_index.less';
+@import 'tokens/_index.less';
 
 // CSS Layers for organized cascade
 @layer reset, base;
