+++ apps/ratio-ui/src/styles/themes/artist.less
@@ -190,9 +190,7 @@
     transform: rotate(0deg);
     border-radius: 0.9375rem;
     overflow: hidden;
-    box-shadow:
-      0.25rem 0.25rem 0.75rem rgba(44, 62, 80, 0.15),
-      0.125rem 0.125rem 0rem rgba(44, 62, 80, 0.2);
+    box-shadow: @artist-shadow-table-strong;
     width: 100%;
     color: var(--text-primary);
 
@@ -202,31 +200,19 @@
       border-bottom: @artist-border-width @artist-border-style @artist-border-color;
 
       &:first-child {
-        .artist-table-header('🏷️ ', 
-          linear-gradient(180deg, #a3d2ff 0%, #baddff 33%, #d1e7ff 66%, #e8f2ff 100%),
-          linear-gradient(180deg, #2f4157 0%, #2d3e52 33%, #2b3a4e 66%, #293649 100%)
-        );
+        .table-header-with-icon('🏷️ ');
       }
 
       &:nth-child(2) {
-        .artist-table-header('⚖️ ', 
-          linear-gradient(180deg, #a3d2ff 0%, #baddff 33%, #d1e7ff 66%, #e8f2ff 100%),
-          linear-gradient(180deg, #2f4157 0%, #2d3e52 33%, #2b3a4e 66%, #293649 100%)
-        );
+        .table-header-with-icon('⚖️ ');
       }
 
       &:nth-child(3) {
-        .artist-table-header('💭 ', 
-          linear-gradient(180deg, #a3d2ff 0%, #baddff 33%, #d1e7ff 66%, #e8f2ff 100%),
-          linear-gradient(180deg, #2f4157 0%, #2d3e52 33%, #2b3a4e 66%, #293649 100%)
-        );
+        .table-header-with-icon('💭 ');
       }
 
       &:nth-child(4) {
-        .artist-table-header('👍 ', 
-          linear-gradient(180deg, #a3d2ff 0%, #baddff 33%, #d1e7ff 66%, #e8f2ff 100%),
-          linear-gradient(180deg, #2f4157 0%, #2d3e52 33%, #2b3a4e 66%, #293649 100%)
-        );
+        .table-header-with-icon('👍 ');
       }
     }
 
@@ -303,27 +289,21 @@
     // Action buttons
     .actions .btn {
       margin: 0 0.25rem;
-      box-shadow: 0.125rem 0.125rem 0.25rem rgba(44, 62, 80, 0.3);
+      box-shadow: @artist-shadow-btn;
       border-width: 0.125rem;
 
       &:hover {
         transform: translateY(-0.0625rem) rotate(@artist-rotate-small);
-        box-shadow: 0.1875rem 0.1875rem 0.375rem rgba(44, 62, 80, 0.4);
+        box-shadow: @artist-shadow-hover;
       }
 
       &.circular {
         &.approve.active {
-          background: linear-gradient(135deg, @success-green, #4caf50);
-          .artist-border(@artist-border-color);
-          .artist-transform(@artist-rotate-small, 1.15);
-          box-shadow: 0.25rem 0.25rem 0.75rem rgba(56, 142, 60, 0.4);
+          .circular-active-state(@success-green, @btn-gradient-success);
         }
 
         &.reject.active {
-          background: linear-gradient(135deg, @error-red, #f44336);
-          .artist-border(@artist-border-color);
-          .artist-transform(@artist-rotate-small, 1.15);
-          box-shadow: 0.25rem 0.25rem 0.75rem rgba(211, 47, 47, 0.4);
+          .circular-active-state(@error-red, @btn-gradient-error);
         }
 
         &:disabled {
@@ -342,25 +322,18 @@
   .table-wrapper {
     .artist-border(@artist-border-color);
     border-radius: 0.9375rem;
-    box-shadow: 0.25rem 0.25rem 0.75rem rgba(44, 62, 80, 0.15);
+    box-shadow: @artist-shadow-table;
 
     &.scrollable::-webkit-scrollbar-thumb {
-      background: linear-gradient(135deg, @primary-green, @success-green);
+      background: @btn-gradient-success;
       border-radius: 0.5rem;
     }
   }
 
   // Toggle buttons
   .artist-toggle {
-    background: linear-gradient(135deg, #ff6b6b, #4ecdc4);
-    color: @white;
-    border: 0.1875rem solid @artist-border-color;
+    .artist-button-gradient(@btn-gradient-artist);
     .artist-transform(@artist-rotate-large);
-    box-shadow: 0.1875rem 0.1875rem 0rem @artist-shadow-color-strong;
-
-    &:hover {
-      .artist-transform(@artist-rotate-large, 1.1);
-    }
   }
 
   .theme-toggle {
@@ -375,16 +348,16 @@
     .artist-border(@artist-border-color);
     border-radius: 1rem;
     .artist-transform(@artist-rotate-small);
-    box-shadow: 0.125rem 0.125rem 0.25rem rgba(44, 62, 80, 0.2);
+    box-shadow: @artist-shadow-btn;
     font-weight: bold;
 
     &:hover:not(:disabled) {
       .artist-transform(@artist-rotate-small, 1, -0.125rem);
-      box-shadow: 0.1875rem 0.1875rem 0.375rem rgba(44, 62, 80, 0.3);
+      box-shadow: @artist-shadow-hover;
     }
 
     &.active {
-      background: linear-gradient(135deg, @primary-green, @success-green);
+      background: @btn-gradient-success;
       border-color: @artist-border-color;
       .artist-transform(@artist-rotate-small, 1.05);
     }
