+++ apps/ratio-ui/src/styles/themes/light.less
@@ -25,21 +25,21 @@
   // Status badges in light theme
   .status-badge {
     &.pending {
-      background: rgba(156, 163, 175, 0.1);
-      color: #6b7280;
-      border-color: rgba(156, 163, 175, 0.2);
+      background: @status-pending-bg;
+      color: @status-pending-text;
+      border-color: fade(@status-pending-text, 20%);
     }
 
     &.approved {
-      background: rgba(56, 142, 60, 0.1);
+      background: @status-approved-bg;
       color: @success-green;
-      border-color: rgba(56, 142, 60, 0.2);
+      border-color: fade(@success-green, 20%);
     }
 
     &.rejected {
-      background: rgba(211, 47, 47, 0.1);
+      background: @status-rejected-bg;
       color: @error-red;
-      border-color: rgba(211, 47, 47, 0.2);
+      border-color: fade(@error-red, 20%);
     }
   }
 
@@ -50,7 +50,7 @@
       color: @white;
       border-color: @success-green;
       transform: scale(1.1);
-      box-shadow: 0 0.125rem 0.5rem rgba(56, 142, 60, 0.4);
+      box-shadow: @shadow-success;
     }
 
     &.reject.active {
@@ -58,18 +58,18 @@
       color: @white;
       border-color: @error-red;
       transform: scale(1.1);
-      box-shadow: 0 0.125rem 0.5rem rgba(211, 47, 47, 0.4);
+      box-shadow: @shadow-error;
     }
   }
 
   // Row states in light theme
   .row-approved {
-    background: rgba(56, 142, 60, 0.05) !important;
+    background: @status-approved-bg !important;
     border-left: 0.25rem solid @success-green;
   }
 
   .row-rejected {
-    background: rgba(211, 47, 47, 0.05) !important;
+    background: @status-rejected-bg !important;
     border-left: 0.25rem solid @error-red;
     opacity: 0.7;
   }
