+++ apps/ratio-ui/src/styles/tokens/_index.less
@@ -108,6 +108,34 @@
   --artist-rotate-large-alt: @artist-rotate-large-alt;
   --artist-brush-cursor: @artist-brush-cursor;
 
+  // Enhanced color tokens
+  --success-green-alt: @success-green-alt;
+  --success-green-shadow: @success-green-shadow;
+  --error-red-alt: @error-red-alt;
+  --error-red-shadow: @error-red-shadow;
+  --status-pending-bg: @status-pending-bg;
+  --status-pending-bg-dark: @status-pending-bg-dark;
+  --status-approved-bg: @status-approved-bg;
+  --status-approved-bg-dark: @status-approved-bg-dark;
+  --status-rejected-bg: @status-rejected-bg;
+  --status-rejected-bg-dark: @status-rejected-bg-dark;
+
+  // Gradient tokens
+  --table-header-gradient-light: @table-header-gradient-light;
+  --table-header-gradient-dark: @table-header-gradient-dark;
+  --btn-gradient-success: @btn-gradient-success;
+  --btn-gradient-error: @btn-gradient-error;
+  --btn-gradient-artist: @btn-gradient-artist;
+
+  // Enhanced shadow tokens
+  --artist-shadow-hover: @artist-shadow-hover;
+  --artist-shadow-btn: @artist-shadow-btn;
+  --artist-shadow-table: @artist-shadow-table;
+  --shadow-success: @shadow-success;
+  --shadow-error: @shadow-error;
+  --btn-shadow-active-success: @btn-shadow-active-success;
+  --btn-shadow-active-error: @btn-shadow-active-error;
+
   // Modern color space support
   @supports (color: oklch(0.7 0.15 180)) {
     --primary-green: oklch(0.45 0.15 140);
