+++ apps/ratio-ui/src/styles/tokens/colors.less
@@ -64,6 +64,49 @@
 @artist-shadow-color: rgba(44, 62, 80, 0.2);
 @artist-shadow-color-strong: rgba(44, 62, 80, 0.4);
 
+// Success Green Variations
+@success-green-alt: #4caf50;
+@success-green-shadow: rgba(56, 142, 60, 0.4);
+
+// Error Red Variations
+@error-red-alt: #f44336;
+@error-red-shadow: rgba(211, 47, 47, 0.4);
+
+// Status Badge Alpha Backgrounds
+@status-pending-bg: rgba(156, 163, 175, 0.1);
+@status-pending-bg-dark: rgba(156, 163, 175, 0.15);
+@status-approved-bg: rgba(56, 142, 60, 0.05);
+@status-approved-bg-dark: rgba(56, 142, 60, 0.15);
+@status-rejected-bg: rgba(211, 47, 47, 0.05);
+@status-rejected-bg-dark: rgba(211, 47, 47, 0.15);
+
+// Table Header Gradients
+@table-header-gradient-light: linear-gradient(
+  180deg,
+  #a3d2ff 0%,
+  #baddff 33%,
+  #d1e7ff 66%,
+  #e8f2ff 100%
+);
+@table-header-gradient-dark: linear-gradient(
+  180deg,
+  #2f4157 0%,
+  #2d3e52 33%,
+  #2b3a4e 66%,
+  #293649 100%
+);
+
+// Button Gradients
+@btn-gradient-success: linear-gradient(135deg, @success-green, @success-green-alt);
+@btn-gradient-error: linear-gradient(135deg, @error-red, @error-red-alt);
+@btn-gradient-artist: linear-gradient(135deg, #ff6b6b, #4ecdc4);
+
+// Status Badge Text Colors
+@status-pending-text: #6b7280;
+@status-pending-text-dark: #9ca3af;
+@status-approved-text-dark: #4caf50;
+@status-rejected-text-dark: #f44336;
+
 // Modern Color Space Support (for browsers that support it)
 .modern-colors() when (@supports-oklch = true) {
   --primary-green: oklch(0.45 0.15 140);
