+++ apps/ratio-ui/src/styles/tokens/shadows.less
@@ -59,6 +59,22 @@
   outline-offset: 0.125rem;
 }
 
+// Artist Mode Shadow Variations
+@artist-shadow-hover: 0.1875rem 0.1875rem 0.375rem rgba(44, 62, 80, 0.4);
+@artist-shadow-btn: 0.125rem 0.125rem 0.25rem rgba(44, 62, 80, 0.2);
+@artist-shadow-table: 0.25rem 0.25rem 0.75rem rgba(44, 62, 80, 0.15);
+@artist-shadow-table-strong:
+  0.25rem 0.25rem 0.75rem rgba(44, 62, 80, 0.15),
+  0.125rem 0.125rem 0rem rgba(44, 62, 80, 0.2);
+
+// Status-specific shadows
+@shadow-success: 0.125rem 0.125rem 0.5rem @success-green-shadow;
+@shadow-error: 0.125rem 0.125rem 0.5rem @error-red-shadow;
+
+// Enhanced button shadows
+@btn-shadow-active-success: 0.25rem 0.25rem 0.75rem @success-green-shadow;
+@btn-shadow-active-error: 0.25rem 0.25rem 0.75rem @error-red-shadow;
+
 // Artist border mixin
 .artist-border(@color: @artist-border-color) {
   border: @artist-border-width @artist-border-style @color;
